Allow us to use node_modules for our react components
settings.py
---
I've configured the django-node-assets package here. This handles adding
the node module files to the static build.

package.json & package-lock.json
---
These files specify the dependencies and the exact version that we
need/have installed. The package.lock file is the only file we should
change, the other is autogenerated.

base.html
---
Updated the static import syntax to point to the files in node_modules.
Note we are always adding the *.min.js files. These are minimized for
production use. Aka their size is compressed.