maps, database, and preliminary react components
This commit contains the updated code for the Database link up, the map
and the react components examples working.

====== Database Link Up =====

The connection information for the RDS database is included with this
commit which allows it to connect it with our database. This will enble
the use of the database when testing the features of the website.

The Database uses a MySQL engine and mySQLclient as a dependency.
Without these the database cannot connect. Additionally it is required
to make sure the MySQL installation is up to date

===== React Components =====

Currently there are two example react components which are shown on the
index (home) page of the application. These are place holder components
and show an example of how the components can work on he site. Currently
they are JS files. JSX files are to be tested to determine if they
follow the same rules.

Currently how they work is to create the component you want to use and
then import it into the index.js file and then call the components in
div id tags in the html of the website after a
{% load render_bundle from webpack_loader %} line of code in the body of
the html page.

The location of this Jinja line may need to be moved into the base.html
file as it appears like it needs to go before the body html tags. More
testing needed for this section.

===== Map =====

The map is rendered using @react-google-maps/api. Currently the map is
not interactive and is only functioning as a premliminary ideology for
the maps for the website. Additional research into how to use the map
components from the documentation at
https://react-google-maps-api-docs.netlify.app/ is required to determine
if the route components will work for our needs, but it looks promising.

Commented out code will need to be reviewed before removal as one of the
examples used in the trying to get the maps to work inculde an example
of the destination components which may be needed in the final map.