validation and error handelign for forms
Updated the way the forms data was cleaned. Changed it from being in the
views.py classes and moved it into a function in the forms.py file.

Additionally implemented the form to inform the user what the issues may
be if the form fails. ie required field or incorrect data.

Started implementing the results page for the journey planner. Currently
it extends the journey planner page complete with the form.