time added to form and info added to map.html
A time input has been added to the form and needed to be added to the
forms.py folder and incorporated into the views.py file too.

Information from the databases and the form can now be displayed on a
html file using a context dictionary. The map.html file is extended from
the journeyPlanner.html file and can display and work the map like the
first page. Additionally there is information being displayed from the
database and the form.