implemented forecast and user time
User Time

Added the ability for the user to add a departure time

Forecasting

Set up the ability to request the current weather information from the
database and produce a dictionary instead of an object.

Additionally the users supplied date and time can now be combined into a
datetime and used to request the weather forecast for the nearest
datetime in the database.