map is now working with user date and addresses
===== Map =====

Map will take an origin and destination address and will return a route
based on the time indicated by the user with only buses being displayed.

JSON file should be able to return the stop ids of the departing stops
and arrival stops depending on the number of transfer. At a later stage
we can use this to determine the estimated time.

===== Form =====

Small changes to the form where completed in order to remove the need
for labels and to display holding text instead.