creating login functions
added in a new app in order to process the login and registration
fucntionality for the website.

Login features created through Djangos default user authentication and
registration templates. Changes needed to be made to the settings.py
file in order to allow the login features.

Restructuring was completed to the static file by moving it to the
DublinBus project folder instead of having it in the app folder. The
templates folder should be moved to the project folder too, additional
research needed for this though.