prediction for estimated time
Implemented a set of functions in the view file which will take
information through a JSON link from the googlemaps api and from the
weather databases and use them to predict the estimated time for a full
route to complete.

Currently the route is hard coded, however this will be adjusted to take
the route names from the google maps API to allow each route to be
predicted.

Additionally due to issues with the folder name for the models folder
containing a hyphen I had to rename the file to modelsNew.