Refactoring the static & Template folders
Removed the static files from the applications and made sure all
references of the static file routes to the new folder location in the
project folder.

Added the missing bootstrap.bundle.min.js.map to the JS folder.

Removed the Footer folder in the JS static folder as it was superseded
code.

Relocated the weather icons from the dublinBusHybrid/static/images folder in the
dublinBusHybrid app to the dublinBus/static/images folder in order to
remove the extra static folder.

Relocated the layouts and partials folders from the template from the
application folders to a templates folder in the project folder.

The relocations were completed in order to allow all applications the
ability to access shared layout and partial files without the need for
duplication.

Updated the CovidInfo.html file to be a content block instead of
Covid-info block in order to allow generic styling of the pages content
and removed reference of the block in the base.html document.

Updated the references to the layouts and partials folders in the html
files to point to the correct folder.