Route section estimated time working
The code is working to return the estimated time of a route we have a
model for. Only retunring to the terminal at this time. Additionaly an
issue with the outbound stop information needs to be solved.

Next step is to include error handeling if the route does not have a
model