Restructure to the journey planner functions
Removed the prediction and pickle functions from the views.py file and
put them into pickelModels.py and predictions.py in order to simplify
the code in views.py. By importing these files as modules the functions
are still accessible but a neater representation.

The functions have been refactored too to allow for this change and to
simplify the functions. Some larger functions have been split up into
smaller functions to allow for easier calling and more general code.