Added date calendar, removed requirement date time
Added a calendar functionality to allow the users to pick a date from a
dropdown calendar and made it so the user can only select the next four
days.

Removed the requirement that the user needs to enter the time and
adjusted the various functions looking for this time with error
checking.