added tabs to the journey planner page
Added 4 tabs to the journey planner page in order to declutter the left
panel. The four tabs are form, info, directions, live.

Form
Shows the form fro the journey planner so that someone can resubmit the
form

Info
Displays the date and time of travel, the estimated journey time and the
weather information for the journey

Directions
Displays the google maps directions panel

Live
Will detail the live information for the bus stops.