Favourite addresses functionality
added the ability for a user to add favourite addresses to allow the
user to have a quicker input time on the journey planner page.

Addresses can be updated, deleted and added.

The addresses are saved as a json object in a table on the database with
a one to one connection with the users id in the users table.