changes to weather functions to accommodate new currentweaher table
Due to the naming of the rain_1h column to rain instead of rain_1h in the current weather table the functions for getweather and getInputValues needded to be adjusted. This allowed the code to be refacotred to include an error case if the weather was to be misbehaving and a message to be displayed instead.