address autocomplete, fix hosting bugs, fix weather bug, github user
*****
address autocomplete

autocomplete added by using the placesAPI to allow people enter addresses easier and to include abbreviations as they're typing

*****
Fix hosting bugs

In order to host the application the settings.py authenticated users needed to allow all users

Additionally, the links in the base.html file needed to have "" enclosing them in order for nginx to read them

*****
Fix weather bug

Bug in the weather where the forecast weather main was trying to call the current weather main when it didn't exist.

*****
github user

added initall code for using the github sign in ability. Code needs to be running on the server in order to be tested.