fix bug with getRouteDirection(), settings.py, allowed_hosts, google warning not displaying
getRouteDirection()

Bug where if the stop search returned None it would fail, added a logic statement that if this happened it would return None

***

Settings.py

Added functionality to the .env file where the Allowed_Hosts is controlled based on whether Debug is True or False

***

Google warning not displaying

Fixed bug whereby the google warning was not displaying if the last bus was a predictions bus and not a google bus.
And the previous bus was a google bus.