autocomplete address and lowercase storage
Autocomplete was added to the input baox for the bookmarked addresses which allows users to input the full address into their bookmarks to avoid acronyms.

Additionally a bug with how the keys were stored has been fixed by making sure all keys stored are lower case. Additionally changes to the dublinBusHybrid/views.py file was made so that when searching for a key word the key word entered is converted to lowercase before being used to search. This avoids the issue with mistypes of cases.