Sign up with gitHub to signup page
The git hub login button on the login page works and a comment was removed

However an additional sign up button was required for the github section which would allow users to sign up through github itself. The ability to signup can be technically used on either page but this gives the option to someone who doesn't know this.